\name{HiCfeat-package}
\alias{HiCfeat-package}
\alias{HiCfeat}
\docType{package}
\title{
\packageTitle{HiCfeat}
}
\description{
\packageDescription{HiCfeat}
}
\author{
Raphael Mourad

Maintainer: Raphael Mourad, raphael.mourad@ibcg.biotoul.fr
}
\references{
Raphael Mourad and Olivier Cuvier. Computational identification of genomic features that influence 3D chromatin domain formation, PLOS Computational Biology, 12(5):e1004908.
}
